Refactor lgraph cmake module
