+++ interactive_engine/lgraph/cmake/FindgRPC.cmake
@@ -19,7 +19,7 @@
 #  DEST - directory where the source files will be created
 #  ARGN - .proto files
 #
-function(GRPC_GENERATE_CPP SRCS HDRS PROTO_ROOT DEST)
+function(GRPC_GENERATE_CPP SRCS HDRS PROTO_ROOT BUILD_DIR SRCS_DEST HDRS_DEST)
     if(NOT ARGN)
         message(SEND_ERROR "Error: GRPC_GENERATE_CPP() called without any proto files")
         return()
@@ -57,16 +57,17 @@ function(GRPC_GENERATE_CPP SRCS HDRS PROTO_ROOT DEST)
         get_filename_component(FIL_WE ${FIL} NAME_WE)
         file(RELATIVE_PATH REL ${PROTO_ROOT} ${ABS_PATH})
 
-        list(APPEND ${SRCS} "${DEST}/${REL}/${FIL_WE}.grpc.pb.cc")
-        list(APPEND ${HDRS} "${DEST}/${REL}/${FIL_WE}.grpc.pb.h")
-
         add_custom_command(
-                OUTPUT "${DEST}/${REL}/${FIL_WE}.grpc.pb.cc" "${DEST}/${REL}/${FIL_WE}.grpc.pb.h"
-                COMMAND protobuf::protoc
-                ARGS --grpc_out ${DEST} ${_protobuf_include_path} --plugin=protoc-gen-grpc=${GRPC_CPP_PLUGIN} ${ABS_FIL}
+                OUTPUT "${SRCS_DEST}/${REL}/${FIL_WE}.grpc.pb.cc" "${HDRS_DEST}/${REL}/${FIL_WE}.grpc.pb.h"
                 DEPENDS ${ABS_FIL} protobuf::protoc gRPC::grpc_cpp_plugin
+                COMMAND protobuf::protoc --grpc_out ${BUILD_DIR} ${_protobuf_include_path} --plugin=protoc-gen-grpc=${GRPC_CPP_PLUGIN} ${ABS_FIL}
+                COMMAND cp ${BUILD_DIR}/${REL}/${FIL_WE}.grpc.pb.cc ${SRCS_DEST}/${REL}/
+                COMMAND cp ${BUILD_DIR}/${REL}/${FIL_WE}.grpc.pb.h ${HDRS_DEST}/${REL}/
                 COMMENT "Running C++ gRPC compiler on ${FIL}"
                 VERBATIM)
+
+        list(APPEND ${SRCS} "${SRCS_DEST}/${REL}/${FIL_WE}.grpc.pb.cc")
+        list(APPEND ${HDRS} "${HDRS_DEST}/${REL}/${FIL_WE}.grpc.pb.h")
     endforeach()
 
     set_source_files_properties(${${SRCS}} ${${HDRS}} PROPERTIES GENERATED TRUE)
