+++ interactive_engine/lgraph/include/lgraph/common/logger_info.h
@@ -16,7 +16,7 @@
 
 #pragma once
 
-#include "common/namespace.h"
+#include "lgraph/common/namespace.h"
 
 namespace LGRAPH_NAMESPACE {
