+++ interactive_engine/lgraph/include/lgraph/common/schema.h
@@ -17,9 +17,9 @@
 #pragma once
 
 #include <fstream>
-#include "common/check.h"
-#include "common/types.h"
-#include "model.pb.h"
+#include "lgraph/common/check.h"
+#include "lgraph/common/types.h"
+#include "lgraph/proto/model.pb.h"
 
 namespace LGRAPH_NAMESPACE {
