+++ interactive_engine/lgraph/include/lgraph/common/types.h
@@ -21,7 +21,7 @@
 #include <cstdint>
 #include <limits>
 
-#include "common/namespace.h"
+#include "lgraph/common/namespace.h"
 
 namespace LGRAPH_NAMESPACE {
