+++ interactive_engine/lgraph/include/lgraph/db/store_ffi/store_ffi.h
@@ -17,9 +17,7 @@
 #pragma once
 
 #include <cstdint>
-
-#include "common/namespace.h"
-#include "common/types.h"
+#include "lgraph/common/types.h"
 
 namespace LGRAPH_NAMESPACE {
