+++ interactive_engine/lgraph/include/lgraph/log_track/message.h
@@ -17,7 +17,7 @@
 #pragma once
 
 #include <cppkafka/consumer.h>
-#include "log_track/operation.h"
+#include "lgraph/log_track/operation.h"
 
 namespace LGRAPH_NAMESPACE {
 namespace log_track {
