+++ interactive_engine/lgraph/src/db/property.cc
@@ -14,8 +14,8 @@
  * limitations under the License.
  */
 
-#include "db/property.h"
-#include "store_ffi/store_ffi.h"
+#include "lgraph/db/property.h"
+#include "lgraph/db/store_ffi/store_ffi.h"
 
 namespace LGRAPH_NAMESPACE {
 namespace db {
